8267163: Rename anonymous loader tests to hidden loader tests
