+++ test/hotspot/jtreg/ProblemList-Xcomp.txt
@@ -35,4 +35,4 @@ vmTestbase/vm/mlvm/mixed/stress/regression/b6969574/INDIFY_Test.java 8265295 lin
 
 vmTestbase/nsk/jvmti/scenarios/sampling/SP07/sp07t002/TestDescription.java 8245680 windows-x64
 
-vmTestbase/vm/mlvm/anonloader/stress/oome/heap/Test.java 8273095 generic-all
+vmTestbase/vm/mlvm/hiddenloader/stress/oome/heap/Test.java 8273095 generic-all
