+++ test/hotspot/jtreg/vmTestbase/vm/mlvm/hiddenloader/stress/oome/metaspace/Test.java
@@ -33,21 +33,21 @@
  *          /test/lib
  *
  * @comment build test class and indify classes
- * @build vm.mlvm.anonloader.stress.oome.metaspace.Test
+ * @build vm.mlvm.hiddenloader.stress.oome.metaspace.Test
  * @run driver vm.mlvm.share.IndifiedClassesBuilder
  *
  * @run main/othervm -XX:MaxRAMPercentage=25 -XX:-UseGCOverheadLimit -XX:MetaspaceSize=10m
- *                   -XX:MaxMetaspaceSize=20m vm.mlvm.anonloader.stress.oome.metaspace.Test
+ *                   -XX:MaxMetaspaceSize=20m vm.mlvm.hiddenloader.stress.oome.metaspace.Test
  */
 
-package vm.mlvm.anonloader.stress.oome.metaspace;
+package vm.mlvm.hiddenloader.stress.oome.metaspace;
 
 import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodHandles.Lookup;
 import java.util.List;
 import java.io.IOException;
 
-import vm.mlvm.anonloader.share.AnonkTestee01;
+import vm.mlvm.hiddenloader.share.HiddenkTestee01;
 import vm.mlvm.share.MlvmOOMTest;
 import vm.mlvm.share.MlvmTestExecutor;
 import vm.mlvm.share.Env;
@@ -71,14 +71,14 @@ public class Test extends MlvmOOMTest {
     protected void eatMemory(List<Object> list) {
         byte[] classBytes = null;
         try {
-            classBytes = FileUtils.readClass(AnonkTestee01.class.getName());
+            classBytes = FileUtils.readClass(HiddenkTestee01.class.getName());
         } catch (IOException e) {
             Env.throwAsUncheckedException(e);
         }
         try {
             while (true) {
                 Lookup lookup = MethodHandles.lookup();
-                Lookup ank_lookup = MethodHandles.privateLookupIn(AnonkTestee01.class, lookup);
+                Lookup ank_lookup = MethodHandles.privateLookupIn(HiddenkTestee01.class, lookup);
                 Class<?> c = ank_lookup.defineHiddenClass(classBytes, true).lookupClass();
                 list.add(c.newInstance());
              }
